Move build subtitle into cipher types
